Solution to Streaming Service Subscription Management

+++ Streaming Service Subscription Management@@67877d61dc107486ad5c2752/code.c
@@ -4,7 +4,7 @@
 struct Subscription {
     char user[20];
     char type[10];
-    int cost;
+    float cost;
 };
 
 int main() {
